+++ packages/htmljs/html.js
@@ -15,7 +15,7 @@ HTML.evaluate = function (node, parentComponent) {
   if (node == null) {
     return node;
   } else if (typeof node === 'function') {
-    return node();
+    return HTML.evaluate(node(), parentComponent);
   } else if (node instanceof Array) {
     var result = [];
     for (var i = 0; i < node.length; i++)
@@ -24,7 +24,7 @@ HTML.evaluate = function (node, parentComponent) {
   } else if (typeof node.instantiate === 'function') {
     // component
     var instance = node.instantiate(parentComponent || null);
-    var content = instance.render();
+    var content = instance.render('STATIC');
     return HTML.evaluate(content, instance);
   }  else if (node instanceof HTML.Tag) {
     var newChildren = [];
@@ -58,7 +58,7 @@ var extendAttrs = function (tgt, src, parentComponent) {
 // These attributes are used to extend `attrs` as long as they are non-nully.
 // All attributes are "evaluated," calling functions and instantiating
 // components.
-HTML.evaluateDynamicAttributes = function (attrs, parentComponent) {
+HTML.evaluateAttributes = function (attrs, parentComponent) {
   if (! attrs)
     return attrs;
 
@@ -82,8 +82,8 @@ HTML.evaluateDynamicAttributes = function (attrs, parentComponent) {
   return result;
 };
 
-HTML.Tag.prototype.evaluateDynamicAttributes = function (parentComponent) {
-  return HTML.evaluateDynamicAttributes(this.attrs, parentComponent);
+HTML.Tag.prototype.evaluateAttributes = function (parentComponent) {
+  return HTML.evaluateAttributes(this.attrs, parentComponent);
 };
 
 // Given "P" create the function `HTML.P`.