+++ packages/mongo-livedata/collection.js
@@ -784,7 +784,7 @@ Meteor.Collection.prototype._validatedUpdate = function(
 
   var doc = self._collection.findOne(selector, findOptions);
   if (!doc)  // none satisfied!
-    return;
+    return 0;
 
   var factoriedDoc;
 
@@ -817,7 +817,7 @@ Meteor.Collection.prototype._validatedUpdate = function(
   // avoid races, but since selector is guaranteed to already just be an ID, we
   // don't have to any more.
 
-  self._collection.update.call(
+  return self._collection.update.call(
     self._collection, selector, mutator, options);
 };
 
@@ -847,7 +847,7 @@ Meteor.Collection.prototype._validatedRemove = function(userId, selector) {
 
   var doc = self._collection.findOne(selector, findOptions);
   if (!doc)
-    return;
+    return 0;
 
   // call user validators.
   // Any deny returns true means denied.
@@ -868,5 +868,5 @@ Meteor.Collection.prototype._validatedRemove = function(userId, selector) {
   // Mongo to avoid races, but since selector is guaranteed to already just be
   // an ID, we don't have to any more.
 
-  self._collection.remove.call(self._collection, selector);
+  return self._collection.remove.call(self._collection, selector);
 };