+++ packages/mongo-livedata/mongo_driver.js
@@ -165,7 +165,7 @@ MongoConnection = function (url, options) {
     self._withDb(function (db) {
       dbNameFuture.return(db.databaseName);
     });
-    self._oplogHandle = new OplogHandle(options.oplogUrl, dbNameFuture);
+    self._oplogHandle = new OplogHandle(options.oplogUrl, dbNameFuture.wait());
   }
 };
 
@@ -791,7 +791,7 @@ var SynchronousCursor = function (dbCursor, cursorDescription, options) {
   self._synchronousNextObject = Future.wrap(
     dbCursor.nextObject.bind(dbCursor), 0);
   self._synchronousCount = Future.wrap(dbCursor.count.bind(dbCursor));
-  self._visitedIds = {};
+  self._visitedIds = new LocalCollection._IdMap;
 };
 
 _.extend(SynchronousCursor.prototype, {
@@ -811,9 +811,8 @@ _.extend(SynchronousCursor.prototype, {
         // because we want to maintain O(1) memory usage. And if there isn't _id
         // for some reason (maybe it's the oplog), then we don't do this either.
         // (Be careful to do this for falsey but existing _id, though.)
-        var strId = LocalCollection._idStringify(doc._id);
-        if (self._visitedIds[strId]) continue;
-        self._visitedIds[strId] = true;
+        if (self._visitedIds.has(doc._id)) continue;
+        self._visitedIds.set(doc._id, true);
       }
 
       if (self._transform)
@@ -853,7 +852,7 @@ _.extend(SynchronousCursor.prototype, {
     // known to be synchronous
     self._dbCursor.rewind();
 
-    self._visitedIds = {};
+    self._visitedIds = new LocalCollection._IdMap;
   },
 
   // Mostly usable for tailable cursors.
@@ -879,9 +878,9 @@ _.extend(SynchronousCursor.prototype, {
     if (ordered) {
       return self.fetch();
     } else {
-      var results = {};
+      var results = new LocalCollection._IdMap;
       self.forEach(function (doc) {
-        results[doc._id] = doc;
+        results.set(doc._id, doc);
       });
       return results;
     }