+++ packages/templating/plugin/html_scanner.js
@@ -152,8 +152,7 @@ html_scanner = {
         if (! name)
           throwParseError("Template has no 'name' attribute");
 
-        // XXX ideally this would read from spacebars-compiler:builtInComponents
-        if (name === "content")
+        if (Spacebars.isReservedName(name))
           throwParseError("Template can't be named \"" + name + "\"");
 
         var renderFuncCode = Spacebars.compile(